# api-interop-layer/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the weather.gov interoperability layer",
        "title": "Weather.gov API Interop Layer",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
        "/county/{fips}": {
            "get": {
                "description": "Get county details, risk overview, and alerts",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "County"
                ],
                "summary": "Get county data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "County FIPS",
                        "name": "fips",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/gridpoints/{wfo}/{x}/{y}/forecast": {
            "get": {
                "description": "Get daily forecast for a grid point",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forecast"
                ],
                "summary": "Get forecast",
                "parameters": [
                    {
                        "type": "string",
                        "description": "WFO",
                        "name": "wfo",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Grid X",
                        "name": "x",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Grid Y",
                        "name": "y",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.ForecastResult"
                        }
                    }
                }
            }
        },
        "/meta/alerts": {
            "get": {
                "description": "Get metadata about alert types",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Meta"
                ],
                "summary": "Get alerts metadata",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/point/{lat}/{lon}": {
            "get": {
                "description": "Get weather data for a specific point including grid, place, alerts, forecast, obs, and radar metadata",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Point"
                ],
                "summary": "Get point data",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Latitude",
                        "name": "lat",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Longitude",
                        "name": "lon",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/products/{productId}": {
            "get": {
                "description": "Get a specific text product by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "productId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/radar/{lat}/{lon}": {
            "get": {
                "description": "Get radar metadata for a point",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Radar"
                ],
                "summary": "Get radar metadata",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Latitude",
                        "name": "lat",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Longitude",
                        "name": "lon",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/risk-overview/{placeId}": {
            "get": {
                "description": "Get risk overview for a place (FIPS or State)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Risk"
                ],
                "summary": "Get risk overview",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Place ID (FIPS or State Abbrev)",
                        "name": "placeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "data.DayPeriod": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Converted properties",
                    "type": "object",
                    "additionalProperties": true
                },
                "end": {
                    "type": "string"
                },
                "isDaytime": {
                    "type": "boolean"
                },
                "isOvernight": {
                    "type": "boolean"
                },
                "start": {
                    "type": "string"
                }
            }
        },
        "data.DayQPF": {
            "type": "object",
            "properties": {
                "hasIce": {
                    "type": "boolean"
                },
                "hasQPF": {
                    "type": "boolean"
                },
                "hasSnow": {
                    "type": "boolean"
                },
                "periods": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": true
                    }
                }
            }
        },
        "data.ForecastDailyResult": {
            "type": "object",
            "properties": {
                "days": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/data.ForecastDay"
                    }
                },
                "elevation": {},
                "error": {
                    "type": "boolean"
                },
                "generatedAt": {
                    "type": "string"
                },
                "updateTime": {
                    "type": "string"
                },
                "validTimes": {
                    "type": "string"
                }
            }
        },
        "data.ForecastDay": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "string"
                },
                "hours": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/data.HourData"
                    }
                },
                "maxPop": {
                    "type": "integer"
                },
                "periods": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/data.DayPeriod"
                    }
                },
                "qpf": {
                    "$ref": "#/definitions/data.DayQPF"
                },
                "start": {
                    "type": "string"
                }
            }
        },
        "data.ForecastResult": {
            "type": "object",
            "properties": {
                "daily": {
                    "$ref": "#/definitions/data.ForecastDailyResult"
                },
                "error": {
                    "type": "boolean"
                },
                "gridData": {
                    "$ref": "#/definitions/data.GridpointResult"
                }
            }
        },
        "data.GridpointResult": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean"
                },
                "geometry": {},
                "qpf": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": true
                    }
                }
            }
        },
        "data.HourData": {
            "type": "object",
            "properties": {
                "properties": {
                    "description": "Properties like 'temperature', 'windSpeed', etc.\nIn TS they put everything on the object.",
                    "type": "object",
                    "additionalProperties": true
                },
                "time": {
                    "type": "string"
                }
            }
        }
    }
}